Move PerOrderIdMutex to util
